10.3.Modular CSS{The Powerful class Attribute Syntaxes}
The Powerful class Attribute Syntaxes
In fact, Vue has special treatment for the class attribute: instead of setting it to a string, you can also pass it an array. Now we can include $style.sidebar and the two static classes inside quotes: p-3 and mb-5.
That should do it! Back at the browser... much better.
So... that's modular CSS! We're going to use it throughout the project and we'll learn a couple more tricks along the way.
You may have noticed that, when you look at the DOM, it's not super clear which component the sidebar class is coming from: that random string doesn't tell us much. That's not a huge deal, but with a little config, we can make this friendlier in dev mode.