10.4.Modular CSS{Controlling the "ident" (modular CSS Class Name)}
Controlling the "ident" (modular CSS Class Name)
Open up webpack.config.js. It doesn't matter where, but down here after enableVueLoader(), I'm going to paste in some code. You can get this from the code block on this page.
I admit, this is a bit low-level. Inside of Webpack, the css-loader is what's responsible for understanding and processing styles. This localIdentName is how the random string is generated when using modular CSS. This tell it to use the component name, then the class name - like sidebar - and then a random hash. And we're only doing this in dev mode because when we build for production, we don't care what our class names look like.

 Tip

Actually, the [name] part will be the filename of the Vue component, not its name.

To make this take effect, at your terminal, hit Ctrl+C to stop Encore and then restart it:

yarn watch
Once that finishes, I'll move back to my browser. Refresh the CSS file first. Nice! The class name is products_sidebar_ random string. And when we try the real page, it works too.
Next, oh, we're going to try something that I'm so excited about. It's called "hot module replacement"... which is a pretty cool-sounding name for something even cooler: the ability to make a change to our Vue code and see it in our browser without - wait for it - reloading the page. Woh.