12.2.HMR: See Changes without Reloading{Disabling CSS Extraction}
Disabling CSS Extraction
But... there's one tiny problem. Don't worry it's not a big deal. Back in our editor, at the bottom, let's make a CSS change: I'm pretty sure a designer just told me that the hover background should be pink.
This time, back on the browser, hmm: the style did not update. But if we refresh, it is there. HMR isn't working for styles.