13.1.Organizing into more Components{Creating the Catalog Component}
One of the huge benefits of Vue is being able to organize your DOM elements and their behavior into smaller pieces. Our app is still pretty simple but you can already see that our template is getting a bit crowded. Could we break this into sub-components?
Let's see: this really looks like two parts: a sidebar and a main area that will eventually list products - I'm going to call that the "catalog". Ok! Let's create two new components to keep things organized.

Creating the Catalog Component
Start with the catalog. Inside the components/ directory - because this will technically be a component that we could re-use, create a new file called catalog.vue.
Then we always start the same - I love when things are boring! Add a <template> tag - with an empty <div></div> to start - and a <script> tag. The minimum we need here is export default an object. To help debugging, we always include at least a name key that identifies this component.