13.2.Organizing into more Components{Creating the Catalog Component}
Nice start team! Let's go grab the parts that we want to move here. Let's see: I want the Products title, the area where that will list products and the legend. Basically, I want to move this entire col-xs-12 div. But... I won't exactly do that. It's up to you, but I think the col-xs-12 belongs in this template because it defines the "layout" for products page. In theory, the catalog component could be embedded into any area on the site, so I won't put this element in there.
Instead, copy the 3 divs inside of this element. Back in catalog.vue, I'm going to keep the empty div. Why? One of the rules of Vue is that you must have a single outer element in each component. If we deleted the div, we would have three outer elements and... you won't be friends with Vue anymore! There is a way to fix this in Vue 3 - called Fragments - and you can even use a a fragments plugin for Vue 2 if you really want to avoid this. But we'll keep the extra div.
In the template, we're using <legend-component /> so we need to import that just like we did before - import LegendComponent from ./legend - and add a components option with that inside.
Perfect: LegendComponent down here, allows us to use legend-component in the template.
The last thing we need is our data: this is the text that we pass to LegendComponent. Copy the data() function from products.vue and paste it here.
Now we could have kept this data key in products and passed it into catalog as a prop. We're going to talk more later about where a piece of data should live and why. But don't worry about that yet.