13.4.Organizing into more Components{Cleaning up the Parent Component}
Cleaning up the Parent Component
I think we're ready! In products.vue, this is going to be so satisfying. Remove the old import and replace it with import Catalog from and go up one level ../components/catalog. Copy that and also import Sidebar from sidebar. Add both of these to the components option to make them available in the template.
Ready to delete some code? Remove all the sidebar markup and instead say: <sidebar />. Do the same for the 3 catalog divs: just <catalog />.
In the component itself, if you look at data(), the legend key is no longer used directly in this component... so we can delete the whole function. Also remove the style tag.
Wow. Look at that! Down to about 25 lines of code! Deciding when a component should be split into smaller components isn't a science. In this case, none of the code we removed was complex, but splitting it allowed us to organize a lot of HTML and styles. I can think more clearly now.
Let's see if it works! Thanks to the dev-server that's running in hot mode, we don't even need to refresh: it does work. But... I'll refresh just to prove it.
Next: as our app grows, there will be more and more directories and paths to keep track of, like going to ../components or ../../scss. That's not a huge deal, but we can add a shortcut to make life easier.