14.2.Aliases{ Using the Alias }
Using the Alias
Start in products.vue. Instead of ../, which gets us up to the js/ directory, we can say @/components/catalog... because @ is an alias to the same directory.
The nice thing is that if we move our code to a different directory, this path will keep working: @ always points to the js/ folder.
We don't have to use this everywhere, but let's update a few other spots, like catalog.vue. Same thing: @/components/legend.
And then in sidebar.vue, it's a bit different. Down in the style tag, we can use the styles alias. But when you're inside CSS code and want to use an alias, you need one extra thing: a ~ prefix. So in this case, ~styles/components.
Open products.js. Instead of ./pages, we can say @/pages.
And one more in app.js. To load the CSS file, we can say styles/ and then we don't the scss directory.
But... maybe you were expecting me to say ~styles like we did earlier? Here's the deal: when you're inside of a JavaScript file, you can just use the word styles even if you referring to a CSS file. The ~ thing is only needed when you're doing the import from inside of CSS itself, like in the style tag.
So... those are Webpack aliases. If you love them, great! If you think they're some sort of strange sorcery, don't use them. The @ alias is common in the Vue world. So at the very least, if you see Vue code importing @/something, now you'll understand the dark magic that makes this work.
Next: let's see our second custom Vue syntax: the v-for directive for looping.