15.1.Looping with v-for{ Creating categories Data }
Let's start to make our app a bit more dynamic. See these categories on the sidebar? They are 100% hardcoded in the template. Boring!
Eventually, when the page loads, we'll make an AJAX request to dynamically load the real categories from our API. This means that the categories will be empty at first and then will change to be the real categories as soon as that AJAX call finishes. And so technically, the categories are something that will change during the lifetime of our Vue app. And anything that changes must live as a key on data.

Creating categories Data
Cool! Let's add our first piece of data to sidebar: data(), then return an object with categories: set to an array. We're going to worry about the AJAX call in a few minutes. For now, let's set the initial data to some hardcoded categories. Each category can look however we want - how about an object with a name property set to our top-selling category - "Dot matrix printers" and a link property set to # for now.
Copy this and create a second category. Set the second category to something just as modern: "Iomega Zip Drives". If you don't know what that is... you're definitely younger than I am.