15.2.Looping with v-for{ Using v-for }
Using v-for
Now that our component has a data called categories, we have a variable called categories in the template! But... we can't just render it with {{ categories }}. Nope, we need to loop over the categories.
Like Twig and its {% for or {% if tags, Vue has a number of custom syntaxes, which are called directives. We'll see the most important ones in this tutorial.
And while some of the Vue directives work a lot like Twig tags, they look a bit different. The directive for looping is called v-for, but instead of being a standalone tag, we put it right on the element that we want to loop.
Check it out: I'll split the <li> onto multiple lines for readability, and then say v-for="category in categories".
That totally custom syntax will loop over the categories data and make a new variable called category available inside the li.
Now... life is easy! For the href="", remember: each category is an object with name and link properties. So we basically want to say category.href. I mean, category.link! I'll catch that mistake in a minute!
But... this won't work yet because it would literally print that string. To make it dynamic, use :href.
Now the contents of the attribute are JavaScript, and category.link is perfectly valid JavaScript.
Below, we can print the name with {{ category.name }}.
That's it! I'll remove the other hardcoded li.
Let's go check it out! When we move over, thanks to our dev-server in hot mode, we don't even need to refresh! The "All Products" is hardcoded, but the two categories below this are dynamic! Oh, but let me fix my mistake from earlier: use category.link because the data has a link property. That looks better.
Back in my editor, ESLint is, once again, mad at me! And that's usually a good hint that I'm doing something silly! It says:

  Elements in iteration expected to have v-bind:key.

That's a fancy way of saying that whenever you use v-for you're supposed to give that element a key attribute, which is any unique identifier for each item in the loop.
Obviously... our code works without it. The problem comes if we updated the categories data. Without a key attribute, Vue has a hard time figuring out which category updated... and which element to re-render.