15.2.Looping with v-for{ The Purpose of the key Attribute; The Directives API Docs; The custom key Attribute }
The Purpose of the key Attribute
Without a key attribute, Vue has a hard time figuring out which category updated... and which element to re-render.
To help Vue out, after v-for, always add a key="". Normally you'll set this to something like category.id - some unique key for each item. In this case, because we're looping over an array we invented, we can use the Array index.
To get access to the array index, change the v-for to category, index. Now say key="index". But... once again... this isn't quite what we want: this would set the key attribute to the string index. To make it dynamic, change it to :key="index".
Hey! Now we even get autocomplete!
Back on the browser, we won't see any difference, but if we updated the categories, Vue would re-render perfectly.

The Directives API Docs
We've now seen two Vue directives. Google for "Vue API" to find a page called Vue.component - API. I love this page! It... well... talks about pretty much everything that exists in Vue.
On the left, I'm going to scroll past tons of things to a section called "Directives". Notice that there aren't that many directives in Vue - about 15 or so. And we've already seen both v-bind and v-for. We'll talk about the other important directives later. For each one, you can get info about how it works.

The custom key Attribute
Oh, and there's one tiny detail I want to mention before we keep going. Back on the site, inspect element on the li. It has class="nav-item", but it does not have a key attribute?
There's some magic going on. Normally, if you add foo="bar" to an element, that element - of course - will have a foo attribute! We know that the true purpose of the key attribute was to help Vue internally... but shouldn't that also cause our element to have a key attribute?
Back on the docs, right below the Directives section is another one called "Special Attributes" and you see that key is one of them. Basically, if you add an attribute in Vue, it will render as an attribute... unless it's one of these special attributes. It's not really an important detail, but if you were wondering why everything other than key is rendered as an attribute, this is your answer. Vue is hiding it, because it knows it's internal.
Next: Vue is more than data and re-rendering: it's also about responding to user interaction. Let's add our first behavior: a link that will collapse and expand the sidebar.