16.1.v-on & methods: User Interaction{Adding the collapsed Data}
We've talked a lot about data, props and using those in a template. But what we haven't done yet is add any behavior. So here's our next big goal: add a link to the bottom of the sidebar that, when the user clicks it, will collapse or expand the sidebar.

Adding the collapsed Data
Cool! But... where should we start? Let's think about it: in the sidebar template, we're going to need to know whether or not we are currently collapsed or expanded... because we will probably need to render different classes or styles to make that happen. And since this "is collapsed" information is something that will change while our Vue app is running, it needs to live in data.
Ok! Inside data(), add a new property called, how about, collapsed. Set it false so that the component starts not collapsed.
Back on the browser, it doesn't do anything yet, but we can see the new data.
Let's update the template to use this. Add a style attribute. Basically, if collapsed is true, we'll set a really small width. Of course we can't just start referencing the collapsed variable right now because we need to add the : in front of :style to make it dynamic.
To set the width, we could put width: in quotes then end quote, plus, and some dynamic logic that uses the collapsed variable. But... yuck! Because the style attribute can get complex, instead of creating a long string of styles, Vue allows us to set this to an object with a key for each style, like width: '500px' and margin: '10px'. In our case, I'll use the ternary syntax: if collapsed is true, set the width.
Ok! We have a collapsed data and we're using it in the template. Testing time! In the Vue dev tools, click on Sidebar, change collapsed to true and... oh that looks awful! We'll clean that up soon. But it is working: the width dynamically changes.