16.2.v-on & methods: User Interaction{Adding the Button}
Adding the Button
Of course, we're not going to expect users to change the collapsed data via their Vue dev tools. Nope: they'll need a button that will change this state.
Back on the template, let's see... after the <ul>, I'll add an <hr>, a div with some positioning classes and a button with class="btn btn-secondary".
Nothing interesting yet! For the button text, if we're currently collapsed, use >>, else, use << Collapse.
Oh ESLint is mad! Hmm, it thinks that the < sign is me trying to open an HTML tag! I could escape this and use &lt; - that's probably a great solution! But as you'll see, Vue will escape this for us.
Back on the browser, I'll inspect the new button, and right click to "Edit as HTML". Yep! Vue automatically escapes any text that you render. That < becomes &lt; automatically.