16.3.v-on & methods: User Interaction{Hello v-text}
Hello v-text
Oh, and, by the way, when you have an element and the only thing inside of it is some dynamic text - like our button - there is one other way to render that text. Copy our dynamic code, add a new attribute called v-text, set it to our dynamic code... and delete the curly braces.
Now Vue is happy. v-text is the third Vue directive that we've seen, after v-bind and v-for. It's not a particularly important one... it's just an alternative way to set the "text" of an element and... it's a tiny bit faster. I mostly just wanted you to see it.