16.4.v-on & methods: User Interaction{ Adding a new Method; OnClick with v-on }
Adding a new Method
Let's get back to the real goal: when the user clicks this button, we need to do something! Specifically, we need to change the collapsed data.
We know that the variables that we have access to in the template are the keys from data and props... though we don't have any props in this component. It turns out that there are a couple other ways to add stuff to your template. One allows you to add functions.
Check it out: it doesn't matter where... but I'll do it after data(), add a methods: option set to an object. Create a function called toggleCollapsed() and, to start, just say console.log('CLICKED! ').
The idea is that, "on click" of this button, we will tell Vue to call this method.

OnClick with v-on
How... do we do that? By using our fourth directive - and a very, very important one. It's called v-on. Inside the button element, add v-on: and then the name of the normal DOM event that you want to listen to, so: click. Set this to the name of the method that should be called on click: toggleCollapsed.
Yep, the toggleCollapsed function is now available in the template because we added it to the methods options. PhpStorm even lets you Ctrl or Cmd click to jump to it!
So... let's try it! Back on your browser, click and... then scroll down. These errors up here were from when we were in the middle of working. And... yes! There's our log!