16.6.v-on & methods: User Interaction{ Updating the State }
Updating the State
Ok: we have a button and when we click it, it calls the toggleCollapsed method. The final step is to change the collapsed state so that the template updates. How? this.collapsed = ! this.collapsed.
If you were eagerly waiting for some complex & impressive code to change the data, I'm sorry to disappoint you.
The secret to this working is that any data keys are accessible as a property on this. In the next chapter, we'll look deeper into how that works. But for now, it feels good: we change the value of this.collapsed and... apparently, Vue will re-render.
Sound too good to be true? Let's try it! Click the button. It works! Over on the Vue dev tools, on the Sidebar component, we can watch the collapsed state change.
And... congratulations! We've now talked about the core parts of Vue. Seriously! To get this working, we added a collapsed data, used that in the template, and, on click, called a toggleCollapsed method that we created... which changed the collapsed data. Vue was then smart enough to re-render the template.
We'll see this basic flow over and over again.
But before we do, I want to talk more about how the this variable works in Vue. It's at the heart of Vue's magic. And if we an understand it, we'll be just a little bit closer to being unstoppable. That's next.