17.2.Magic "this" & its Properties{   Variables in the Template are called on this}
Let's try something else: print definitelyNotARealProperty. Back on the browser... the error is wonderful:

  Property or method definitelyNotARealProperty is not defined on the instance but was referenced during render.

Vue is literally saying: this is not a property on the instance! And the way you add something to the instance is by defining it under data, props, methods or the computed option that we'll talk about later. I love it!
Next, Vue has one other piece of magic I want to explore called reactivity. It's all about how Vue is smart enough to re-render a template at the instant that we change a simple piece of data.