18.1.Reactivity{   Data & Props are Getter Properties; Proxy Objects }
We now understand that we pass Vue a set of options and it takes all the keys in data, props and methods and adds those onto the Vue instance. We can see this in our console thanks to the console.log(this) that we added: this has categories, collapsed and testProp properties plus a toggleCollapsed method. Vue effectively copies these onto the instance so that we can say things like this.collapsed.
But there's still a mystery: when we change a data key - like this.collapsed = ! this.collapsed - Vue instantly re-renders the template. How does that happen? Think about it: if collapsed is just a simple property, then how is Vue aware that we changed it?

Data & Props are Getter Properties
The answer is... magic! Sorcery! Potions! Ok... really it's just clever JavaScript. Look back at the categories and collapsed properties on the console. Notice that, instead of showing the value, it shows (...) and says "invoke property getter" when I hover. If you click that, then it shows the value.
It turns out that categories, collapsed and also testProp are not real properties on the object! Scandal! They're "getter" properties. If you scroll near the bottom of the log, you see get categories and get collapsed. What you're seeing is a special feature of JavaScript where you can make an object look like it has a property, even if it really doesn't. There is no property called categories, but we are allowed to reference this.categories thanks to this get categories function. When we say this.categories or this.collapsed, it calls this proxyGetter, which is some low-level Vue function. And when we set that property - like this.collapsed = something, it will call this proxySetter function.
The point is: Vue makes our data accessible, but indirectly via these getter and setter functions. It does that so that it can hook into our calls and re-render. When we say this.collapsed =, that calls proxySetter, which updates the data and tell Vue that it needs to re-render any affected components.
And... this is great! We get to run around just saying this.collapsed = not even realizing that Vue is intercepting that call and intelligently re-rending.

Proxy Objects
By the way, in Vue 3, this magic is done by something called a Proxy. A Proxy is a native JavaScript object - it's a feature built into the language itself, not invented by Vue. With a Proxy, you can wrap another object and intercept all method calls, property gets and property sets. The result is the same... it's just a bit of a cleaner way to get the job done. By the way, this whole idea of hooking into us changing data and then automatically re-rendering any affected templates has a cool name: reactivity. Oooo.