18.2.Reactivity{  Reactivity in Arrays; Deep Reactivity }
Reactivity in Arrays
Reactivity gets even more amazing when you look at the categories data. Back in created, let's also log this.categories for simplicity. Back on the console... find that log - it's on the right - and expand it.
Check it out: it's a normal Array with 0 and 1 keys. So... if we dynamically added a third item to this array, would Vue know to re-render? Yep!
In Vue 3, instead of an array, this would be a Proxy around an array, which would let Vue hook into the new item being added. How is it done in Vue 2? Check out this __proto__ key. This is a fancy place where we can see all the methods that exist on this Array, which is an object in JavaScript. So you can call .pop, .push() or any of these. But check it out: each is set to a function called mutator(). That is not a native JavaScript function: it comes from Vue!
Yea, Vue has replaced all of the Array functions with their own functions! If we push() a new item onto the array, it will call this mutator() function, which will push the item on the array and trigger any re-rendering. It's... pretty crazy. And actually, because of how this is done in Vue 2, there are a few edge case ways that you update an Array in a way that Vue cannot detect. The most common by far is if you set an item directly to a specific index. That's not a problem in Vue 3 thanks to the Proxy object.

Deep Reactivity
But wait, there's more! Check out one of the individual category objects. Remember, in data, each category has name and link properties. Guess what? Once Vue loads the data, these are not real properties anymore: it says "invoke getter". Vue creates an object that looks like the data we created, but instead of having real name and link properties, it adds getters and setters for them - the same reactiveGetter and reactiveSetter functions we saw earlier. So if we changed the name property of the 0 index category, that will update that property and trigger any re-rendering needed.
Phew! So this is the real magic of Vue, and one of the things that sets it apart from React. In React, you handle data - or state - a bit more manually, but with less magic. It's always a tradeoff. But if you can understand how Vue's magic works on a high level, it will go a long way to helping you do great stuff.