19.1.v-if, v-show and Conditional Classes{v-if directive}
Our collapsing sidebar works, but yikes, it looks terrible! We need to fix this! Let's hide the text and links entirely when the sidebar is in its collapsed state.

v-if directive
Look at sidebar.vue. In the template, we want to hide the <h5>, the <ul> and the <hr />: basically everything except the button.
In Vue, there are two ways to do this and both of them come in the form of directives. To make this work, I'm going to add a new <div> to wrap all the elements that we need to hide.
The first directive is v-if. We can write v-if="" and, inside, use a JavaScript expression. We'll say !collapsed. So: show this if we are not collapsed. If we go over to our browser, we don't even need to refresh because we're using hot module replacement. And... it works! Much better!
The key thing here, if you inspect the HTML, is that with v-if, the elements are removed entirely. So if we click the button again, the elements come back, and if we collapse, they're gone!